+++ test/serialization.cpp
@@ -282,27 +282,33 @@ BOOST_AUTO_TEST_CASE(test_collision_data) {
   Sphere sphere(1.);
   Box box(Vec3s::Ones());
 
-  auto contact_remap = [&](Contact& c) { c.remap(&sphere, &box); };
+  // testing serialization with valid pointers
+  // we need to add a post load hook to remap the contact's internal pointers to
+  // the shapes
+  Contact contact0(&sphere, &box, 1, 2, Vec3s::Ones(), Vec3s::Zero(), -10.);
+  test_serialization(contact0,
+                     [&](Contact& c) { c.resolveReferences({&sphere, &box}); });
 
-  Contact contact(&sphere, &box, 1, 2, Vec3s::Ones(), Vec3s::Zero(), -10.);
-  test_serialization(contact, contact_remap);
+  // testing serialization with NULL pointers
+  Contact contact1(NULL, NULL, 1, 2, Vec3s::Ones(), Vec3s::Zero(), -10.);
+  test_serialization(contact1);
+
+  Contact contact2;
+  test_serialization(contact2);
 
   CollisionRequest collision_request(CONTACT, 10);
   test_serialization(collision_request);
 
   CollisionResult collision_result;
-  collision_result.addContact(contact);
-  collision_result.addContact(contact);
+  collision_result.addContact(contact0);
+  collision_result.addContact(contact1);
+  collision_result.addContact(contact2);
   collision_result.distance_lower_bound = Scalar(0.1);
   collision_result.normal.setOnes();
   collision_result.nearest_points[0].setRandom();
   collision_result.nearest_points[1].setRandom();
   test_serialization(collision_result, [&](CollisionResult& cr) {
-    for (size_t i = 0; i < cr.numContacts(); ++i) {
-      Contact c = cr.getContact(i);
-      contact_remap(c);
-      cr.setContact(i, c);
-    }
+    cr.resolveReferences({{&sphere, &box}, {NULL, NULL}, {NULL, NULL}});
   });
 
   DistanceRequest distance_request(true, 1., 2.);
@@ -315,8 +321,9 @@ BOOST_AUTO_TEST_CASE(test_collision_data) {
   distance_result.nearest_points[1].setRandom();
   distance_result.o1 = &sphere;
   distance_result.o2 = &box;
-  test_serialization(distance_result,
-                     [&](DistanceResult& dr) { dr.remap(&sphere, &box); });
+  test_serialization(distance_result, [&](DistanceResult& dr) {
+    dr.resolveReferences({&sphere, &box});
+  });
 
   {
     // Serializing contact patches.